Refactor `ShowInWindow` and move utility functions to a shared file
- Updated `showInWindow` to accept an optional `NSApplication` parameter for better testability and flexibility.
- Extracted common utility functions (e.g., `readImage`, `hash`, `HashAlgorithm`) and moved them to a new shared file `common-functions.swift`.
- Cleaned up `ShowInTerminalTest` by removing redundant code and aligning with the shared utility functions.
